==> 03 Scripts/5. data consistency checks.ipynb <==
Table of Content 
1.Importing libraries 
2.importing data 
3.practising with test data frame
4.identifying missing values
5.creating data frames with the missing values
6.finding mean and median and imputing missing values if possible
7.comparing the number of rows in the current dataframe with the numbers in the subset
8.DUPLICATES
9.exporting the clean data sets
10.EXERCISE 4.5
 10.1 Quick stats check (Task 2)
 10.2 Checking the df for mixed data columns  (Task 3)
 10.3 Missing values (TASK 5)
 10.4 Adressing missing values (Task 6)
 10.5 looking for full duplicates (Task 7)
 10.6 Exporting the Data 

## Importing libraries 

import pandas as pd
import numpy as p
import numpy as np
import os 

## importing data 

#importing products.csv 
df_prods=pd.read_csv(r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis/02 Data/Original Data/products.csv', index_col = False)

#importing orders_wrangled.csv
df_ords= pd.read_csv(r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis/02 Data/Prepared Data/orders_wrangled.csv', index_col = False)

## practising with test data frame 

#create a test data frame 
df_test=pd.DataFrame()
df_test

#create a mix type column
df_test['mix']= ['a','b',1, True ]
df_test

#checking whether a dataframe contains any mixed-type columns 
for col in df_test.columns.tolist():
  weird = (df_test[[col]].map(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

#changing the data type 
df_test['mix'] = df_test['mix'].astype('str')

## identifying missing values 

# PRODS checking the amount of missing values 
df_prods.isnull().sum()


## creating data frames with the missing values 

#creating a data frame with only the missing values
df_nan_prods=df_prods[df_prods['product_name'].isnull()== True]

#PRODS checking the data frame with missing values 
df_nan_prods

## finding mean and median and imputing missing values if possible 

#PRODS look for the mean of the colums 
df_prods.describe()

## comparing the number of rows in the current dataframe with the number in the subset 
##once the missing rows have been removed

#PRODS rows and columns of df_prods with nan entries 
df_prods.shape 

#PRODS setting the isnull() condition to False instead of True (you want non-missing values in your new dataframe 
#as opposed to missing values).
df_prods_clean= df_prods[df_prods['product_name'].isnull() == False]

# PRODS rows and columns of df_prods after removing nan entries 
df_prods_clean.shape

-> df_prods_clean has 16 rows less then original df_prods, same amount like rows with missing entries! 

## DUPLICATES 

#PRODS look for full duplicates within your prods dataframe:

df_dups_prods = df_prods_clean[df_prods_clean.duplicated()]

Addressing Duplicates

#PRODS checking the current number of rows in the df_prods_clean dataframe so that we can compare the number after 
#removing the duplicates
df_prods_clean.shape

#PRODS create a new dataframe that doesn’t include the duplicates just identified 
#using the drop_duplicates() function:

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

#PRODS checking the number of rows after dropping the duplicates 
df_prods_clean_no_dups.shape 

## exporting the clean data sets 

#creating a path variable 
path= '/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis'

df_prods_clean_no_dups.to_csv(os.path.join(path,'02 Data','Prepared Data', 'products_checked.csv'))

## EXERCISE 4.5 

## Quick stats check  (Task 2)

#whether anything about the data looks off or should be investigated further
df_ords.describe()

-> the first column is called "Unnamed: 0	" Without further information, it is impossible to draw conclusions or interpret insights from the data 

-> with so many entries it could be strange that the maximum of days_since_last_order is 30 days, at least it means that the customers were active the last 30 days without 

-> with an mean of aroun 11 days in the column days_since_last_order, it means that the majority of the orders are not far back in time.

## Checking the df for mixed data columns  (Task 3)

#Check for mixed-type data in your df_ords dataframe
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)
  else:
    print('no mix')
    

# to find out if 7* 'no mix' is right i check ords for the number of columns and rows 
df_ords.shape 

->data frame has 7 columns, so it makes sense that "no mix" appeared 7 times, the function checked every column and printed that they are not mixed columns

#If you find mixed-type data, fix it. The column in question should contain observations of a single data type.

-> no mixed type data 

## Missing values (TASK 5)

# checking for missing values in df_ords dataframe.
df_ords.isnull().sum()


-> there are 206209 missing values in the "days_since_last_order" column of the ords data frame. 
-> a possible explanation would be that the orders are first time orders, therefore there are no recorded orders 
before this first one. 
-> but it could also be a hint for errors, would make sense to have a look at the customer ids oder order ids and check 
if other purchases were made in the past 


## Adressing missing values (Task 6)

#Address the missing values 
#creating a data frame with only the missing values
df_nan_ords=df_ords[df_ords['days_since_last_order'].isnull()== True]

#checking the data frame with missing values 
df_nan_ords

#renaming the new data frame with the missing entries more intuitive 
df_first_time_customers = df_nan_ords

#creating a new data frame without the first time customers 
df_ords_returning_customers = df_ords[df_ords['days_since_last_order'].isnull() == False]

#checking df_ords 
df_ords.shape

#df_ords_returning_customers
df_ords_returning_customers.shape

-> when subtracting the count of columns from df_ords_returning_customers from df_ords I get 206209. 
This is the number of missing entries. Means that the new data frame really is without first time customers. 

# creating a new column 'customer_type' in the original DataFrame 'df_ords'
df_ords['customer_type'] = np.where(df_ords['days_since_last_order'].isnull(), 'new_customer', 'returning_customer')


checking new df_ords

-> for now I wont do anything else with the missing values, because theyr missing makes sense. 

## looking for full duplicates (Task 7)

#look for full duplicates within ords dataframe:
df_dups_ords = df_ords[df_ords.duplicated()]

#checking the current number of rows in the ords dataframe so that we can compare the number after 
#removing the duplicates
df_ords.shape

#create a new dataframe that doesn’t include the duplicates just identified 
#using the drop_duplicates() function:

df_ords_no_dups = df_ords.drop_duplicates()

df_ords_no_dups.shape

-> there were no full duplicates spotted. 


## Exporting the Data 

#creating a path to the main folder 
path: r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis'

#exporting the cleaned df_ords dataframe 
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'df_ords_cleaned_ex4.5.csv'))
