==> 03 Scripts/6.1.combining and exporting data.ipynb <==
Table of Content 
1.import libraries 
2.importing data 
3.checking the dimensions of the data 
4.merging two dataframes
5.Exporting the Data in PKL format 

# Import libraries
import pandas as pd
import numpy as np
import os

## importing data

df_ords_prior=pd.read_csv(r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis/02 Data/Original Data/orders_products_prior.csv', index_col = False)

df_ords =pd.read_csv(r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis/02 Data/Prepared Data/df_ords_checked.csv', index_col = False)

## checking the dimensions of the data 

df_ords_prior's dimensions 

df_ords_prior.shape

df_ords_prior.describe()

df_ords_prior.head()

df_ords's dimensions 

df_ords.shape

df_ords.describe()

df_ords.head()

## merging two dataframes 

#merging the new data frame with the ords data frame 
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id',indicator = True)

#checking for a full match 
df_merged_large['_merge'].value_counts()

#merging 
#( how = 'outer': Merging like this will combine all the observations and show you the real merge rate )
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', how = 'inner',indicator = True)

df_merged_large.shape

df_merged_large.head()

#checking for a full match 
df_merged_large['_merge'].value_counts()

->After using this method to double-check your merge, you can see that you don’t actually have a full match.

## Exporting the Data in PKL format 

path = '/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis'

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
