==> 03 Scripts/6.2 combining and exporting data.ipynb <==
Table of Content 
1.importing libraries
2.importing data 
3.checking the shape of the two data sets
4.thinking of suitable way to combine the two data sets (Task 5)
5.full match check 
6.Exporting Data in PKL format

# importing libraries
import pandas as pd
import numpy as np
import os

## importing data

df_ords_prods_merged =pd.read_pickle(r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis/02 Data/Prepared Data/orders_products_combined.pkl')
df_prods=pd.read_csv(r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis/02 Data/Prepared Data/products_checked.csv', index_col = False)

## checking the shape of the two data sets

df_ords_prods_merged.shape

df_prods.shape 

df_prods.head()

df_ords_prods_merged.head()

## thinking of suitable way to combine the two data sets (Task 5)

1) If there are common columns between the DataFrames that can be used as keys to merge the data, I might use techniques like inner join, outer join, left join, or right join, depending on the output I want ->merged_df = pd.merge(df_a, df_b, on='common_column', how='inner')

2) If the DataFrames have no common columns but have the same structure (same columns) and I want to stack them vertically, I would concatenate them using pd.concat()



#Before merging the columns I will drop some unuseful columns from the data frame in order to save time 
#and reduce the costs of the merging command. Plus it will give a clearer overview about the data.

df_prods = df_prods.drop(columns = ['Unnamed: 0'])

#Before merging the columns I will drop some unuseful columns from the data frame in order to save time 
#and reduce the costs of the merging command. Plus it will give a clearer overview about the data.
df_ords_prods_merged = df_ords_prods_merged.drop(columns = ['Unnamed: 0','Unnamed: 0.1','_merge'])

#Merge df_prods and df_ords_prods_merged as inner join 
df_merged = df_prods.merge(df_ords_prods_merged, on = 'product_id', indicator = True)

## full match check 

#checking for a full match 
df_merged['_merge'].value_counts()

#merging this time with outer join
#( how = 'outer': Merging like this will combine all the observations and show  the real merge rate )
df_merged = df_prods.merge(df_ords_prods_merged, on = 'product_id', how = 'outer',indicator = True)

#checking againfor a full match
df_merged['_merge'].value_counts()

#checking the rows and columns after merging 
df_merged.shape

df_merged.head()

## Exporting Data in PKL format

path = '/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis'
df_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged_2.pkl'))
