==> 03 Scripts/7.Deriving New Variables .ipynb <==
Table of Content
1.Importing libraries 
2.Importing the data 
3.creating a price_label function 
4.checking the data 
5.using Loc Function to create flags (Mid-range product or low-range product)
6.Trying the loc function with the whole data set 
7.finding out on which day most orders are placed
8.creating new column : busiest day
9.EXERCISE 4.7
 9.1 creating new column 'busiest_days' (Task 2)
 9.2 Accuracy check (Task 3)
 9.3 Creating new colum “busiest_period_of_day" (Task 4)
 9.4 checking the results (Task 5)
 9.5 Exporting the data (Task 6)

## Importing libraries

import pandas as pd
import numpy as np
import os

## Importing the data 

ords_prods_merge=pd.read_pickle(r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis/02 Data/Prepared Data/orders_products_merged_2.pkl')
                               

#creating a subset of the first million rows 
df = ords_prods_merge[:1000000]

## creating a price_label function 

# defining the function 
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

#calling the function 
df['price_range'] = df.apply(price_label, axis=1)

## checking the data 

#checking the df
df

# checking rows and columns 
df.shape

#counting the values 
df['price_range'].value_counts(dropna = False)

#checking whats the most expensive prdouct of the subset 
df['prices'].max()

# using Loc Function to create flags (Mid-range product or low-range product)

#creating some conditions 
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

#counting the values 
df['price_range_loc'].value_counts(dropna = False)

#counting the values 
df['price_range_loc'].value_counts(dropna = False)


## Trying the loc function with the whole data set 

ords_prods_merge.loc[ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'


ords_prods_merge.loc[(ords_prods_merge['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

ords_prods_merge.loc[ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

#counting the values 
ords_prods_merge['price_range_loc'].value_counts(dropna = False)


## finding out on which day most orders are placed

#which day do most orders take place? finding this out by printing the frequency of the “orders_day_of_week” column
ords_prods_merge['orders_day_of_week'].value_counts(dropna=False)

## creating new column : busiest day

#use this information to create a new column, “busiest day,” that will contain one of three different values: 
#“Busiest day,” “Least busy,” and “Regularly busy.”

result = []

for value in ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

#checking the first 20 rows 
result[:20]

#combining results with ords_prods_merge dataframe.
#creating a summary column for the busiest day of the week

ords_prods_merge['busiest_day']= result 

#printing the frequency of this new column and cross-checking it with the frequency of “orders_day_of_week” 
ords_prods_merge['busiest_day'].value_counts(dropna=False)

# EXERCISE 4.7 

## creating new column 'busiest_days' (Task 2)

Suppose your clients have changed their minds about the labels you created in your “busiest_day” column. 
Now, they want “Busiest day” to become “Busiest days” (plural). This label should correspond with the two 
busiest days of the week as opposed to the single busiest day. At the same time, they’d also like to know 
the two slowest days. Create a new column for this using a suitable method.

#creating a new loop to label the busiest and least busiest days. saving the results in an empty list
new_result = []

for value in ords_prods_merge["orders_day_of_week"]:
    if value == 0 or value == 1:
        new_result.append("Busiest days")
    elif value == 4 or value == 3:
        new_result.append("Least busy")
    else:
        new_result.append("Regularly busy")

#adding the values to the new created column
ords_prods_merge['busiest_days']= new_result 

#printing the frequency of this new column and cross-checking it with the frequency of “orders_day_of_week” 
ords_prods_merge['busiest_days'].value_counts(dropna=False)

## Accuracy check (Task 3)

#Check the values of this new column for accuracy. 
#adding the values from the new_result:
12916111+ 11864412+7624336

#adding values from old result: 
22416875+ 6204182+ 3783802

## Creating new colum “busiest_period_of_day" (Task 4)

When too many users make Instacart orders at the same time, the app freezes. 
The senior technical officer at Instacart wants you to identify the busiest hours of the day. 
Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” 
Create a new column containing these labels called “busiest_period_of_day.”

#checking the frequency of the orders by hour of the day 
ords_prods_merge['order_hour_of_day'].value_counts(dropna=False)

-> there are 24 hours. dividing it through 8, there are 3 periods. first 8 entries are “Most orders", the next 8 
“Average orders” and the rest "Fewest Orders". 

One approach : 
    
order_frequency = []

for value in ords_prods_merge["order_hour_of_day"]:
    if value == 10 or value == 11 or value == 14 or value == 15 or value ==13 or value == 12 or value ==  16 or value == 9:
        order_frequency.append("Most orders")
    elif value == 17 or value == 8 or value == 18 or value == 19 or value == 20 or value == 7 or value == 21 or value == 22:
        order_frequency.append("Average orders")
    else:
        order_frequency.append("Fewest orders")
        
-> this code works BUT when the underlying data changes, the results will not be valid anymore. 
Therefore I try another aproach, which creates out of the output from "ords_prods_merge['order_hour_of_day'].value_counts(dropna=False)"
a list which is saved in th variable called "hours_by_orders". This way the code will be useful even if the sales 
 will change, because its based on the indexes. 

#I want the values from the left side(hours) from the output from following function : ords_prods_merge['order_hour_of_day'].value_counts(dropna=False)
#as an list. Save it in the variable hours_by_orders
hours_by_orders=ords_prods_merge['order_hour_of_day'].value_counts(dropna=False).index.to_list()

#I take the first 8 values from the list and save it in an variable 
Most_ordered_hours=hours_by_orders[0:8]

#I take the middle 8 values from the list and save it in an variable 
Avg_ordered_hours=hours_by_orders[8:16]

#I take the last 8 values from the list and save it in an variable 
fewest_ordered_hours=hours_by_orders[16:24]

#now i change the code so that it would work even if the unerlying order data would change

order_frequency = []

for value in ords_prods_merge["order_hour_of_day"]:
    if value in Most_ordered_hours:
        order_frequency.append("Most orders")
    elif value in Avg_ordered_hours:
        order_frequency.append("Average orders")
    else:
        order_frequency.append("Fewest orders")

#Create a new column containing these labels called “busiest_period_of_day.”
ords_prods_merge['busiest_period_of_day']= order_frequency

#checking the table with the new column and its entries 
ords_prods_merge.head()

## checking the results (Task 5)

ords_prods_merge['busiest_period_of_day'].value_counts(dropna=False)

#checking the sum of the values
21118071+  9997651+  1289137

## Exporting the data (Task 6)


path = '/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis'

ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge.pkl'))
