==> 03 Scripts/8.  Grouping Data & Aggregating Variables.ipynb <==
Table of Content 
1.importing libaries 
2.importing the data 
3.Task 2 
 3.1 checking the df 
 3.2 find the aggregated mean of the “order_number” column grouped by “department_id” 
4. Task 4 
 4.1 creating a loyalty flag for existing customers using the transform() and loc() functions 
5. Task 5 
 5.1 checking the basic statistics of the product prices for each loyalty category 
6. task 6 
 6.1 Creating a spending flag for each user
7.Task 7
 7.1 determining frequent versus non-frequent customers
8. task 8 
 8.1 Exporting the data 


## Importing libraries


import pandas as pd
import numpy as np
import os

## Importing the data


ords_prods_merge=pd.read_pickle(r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis/02 Data/Prepared Data/ords_prods_merge.pkl')

df=ords_prods_merge

# Task 2

## Checking the df 

df.shape

df.head()

# deleting an unuseful column 
df = df.drop(columns = ['_merge'])

#checking if deleting the column was succesful 
df.shape

## find the aggregated mean of the “order_number” column grouped by “department_id”

#grouping the department_id 
df.groupby('department_id')

#performing a single aggregation 
df.groupby('department_id').agg({'order_number': ['mean']})

-> How do the results for the entire dataframe differ from those of the subset?
the results dont differ too much from those of the subset. Some rows a nearly similar, some vary by some thousand 
(for example row 21 has in the subset a 2000 higher value) 

# Task 4

## creating a loyalty flag for existing customers using the transform() and loc() functions (Task4)

#Split the data into groups based on the “user_id” column.
#Apply the transform() function on the “order_number” column to generate the maximum orders for each user.
#Create a new column, “max_order,” into which you’ll place the results of your aggregation.
df['max_order'] = df.groupby(['user_id'])['order_number'].transform(np.max)


#check out the results using the head() function:
df.head(15)

#Deriving Columns with loc()
df.loc[df['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
df.loc[(df['max_order'] <= 40) & (df['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
df.loc[df['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

#checking the count of the different labels 
df['loyalty_flag'].value_counts(dropna=False) 

#checking first 20 rows of the loyalty flag colum
df['loyalty_flag'].head(20)

## Task 5 

## checking the basic statistics of the product prices for each loyalty category 

#Use the loyalty flag you created and check the basic statistics of the product prices for each loyalty category 
#(Loyal Customer, Regular Customer, and New Customer). What you’re trying to determine is whether the prices of 
#products purchased by loyal customers differ from those purchased by regular or new customers.
loyalty_price_stats = df.groupby('loyalty_flag')['prices'].describe()

print(loyalty_price_stats)

-> The regular customers spend most money when buying products. New customers though have the highest mean, 
followed by regular customers and then loyal customers. 

## Task 6 

## Creating a spending flag for each user

looking at the prices of the items people are buying. Create a spending flag for each user based on the average 
price across all their orders using the following criteria:
If the mean of the prices of products purchased by a user is lower than 10, then flag them as a “Low spender.”
If the mean of the prices of products purchased by a user is higher than or equal to 10, then flag them as a
“High spender.”


# Calculate the mean price of products purchased by each user
df['user_mean_price'] = df.groupby('user_id')['prices'].transform('mean')


df.head(20)

# Create a spending flag column based on mean price criteria using loc
df.loc[df['user_mean_price'] < 10, 'spending_type'] = 'Low spender'
df.loc[df['user_mean_price'] >= 10, 'spending_type'] = 'High spender'

df.head(20)

# Task 7 

## determining frequent versus non-frequent customers

In order to send relevant notifications to users within the app (for instance, asking users if they want to buy the same item again), the Instacart team wants you to determine frequent versus non-frequent customers. 

Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column. The criteria for the flag should be as follows:

If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer.”
If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”
If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”

# Calculate the median order frequency by each user
df['order_frequency'] = df.groupby('user_id')['days_since_last_order'].transform('median')

#checking the result column 
print(df['order_frequency'].head(10))

# Create a frequency flag column based on order_frequency using loc
df.loc[df['order_frequency'] > 20, 'frequency_flag'] = 'Non-frequent customer'
df.loc[df['order_frequency'] >10 & (df['order_frequency']<= 20),'frequency_flag'] = 'Regular customer'
df.loc[df['order_frequency'] <= 10, 'frequency_flag'] = 'Frequent customer'

## Task 8 

## Exporting the data 

path = '/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis'

df.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge_4.8.pkl'))
