==> 03 Scripts/9.1 Data Visualization with Python part_1..ipynb <==
 Table of Content 
1.Importing libaries 
2.importing the data set
3.data wrangeling (Task 4)
4.Completing the fundamental data quality and consistency checks (Task 5)
5.changing data types
6.looking for mixed data type columns in the df
7.Finding missing entries
8.Looking for full duplicates
9.Combining the customer data with the rest of the prepared Instacart data (Task 6) 
10.Exporting the data set

## Importing libaries 

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy


## importing the dataset¶

cust=pd.read_csv(r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis/02 Data/Original Data/customers.csv', index_col = False)

## data wrangeling (Task 4 )

#Wrangle the data so that it follows consistent logic; for example, rename columns with illogical 
#names and drop columns that don’t add anything to your analysis.

cust.head(10)

cust.rename(columns={'STATE': 'State'}, inplace=True)

cust.rename(columns={'date_joined': 'Joining Date'}, inplace=True)

cust.rename(columns={'n_dependants': 'Dependants'}, inplace=True)

cust.rename(columns={'fam_status': 'Family Status'}, inplace=True)

# in case I want to drop a column: 
#cust.drop(columns='column_to_delete', inplace=True)

cust.head(10)

cust.rename(columns={'user_id': 'User Id'}, inplace=True)

cust.rename(columns={'fam_status': 'Family_status'}, inplace=True)

cust.rename(columns={'income': 'Income'}, inplace=True)

cust.rename(columns={'Family_status': 'Family Status'}, inplace=True)

cust.head(10)

## Completing the fundamental data quality and consistency checks (Task 5)


#Complete the fundamental data quality and consistency checks you’ve learned throughout this Achievement; 
#for example, check for and address missing values and duplicates, and convert any mixed-type data.


cust['First Name'] = cust['First Name'].astype('str')

cust.dtypes

## changing data types

#changing the data type "object" to string to make sure it contains no mixed values 

cust['First Name'] = cust['First Name'].astype('str')

cust['Surname'] = cust['Surname'].astype('str')

cust['Gender'] = cust['Gender'].astype('str')

cust['State'] = cust['State'].astype('str')

cust['Joining Date'] = cust['Joining Date'].astype('str')

cust['Family Status'] = cust['Family Status'].astype('str')

cust['User Id'] = cust['User Id'].astype('str')

I decided to not drop any columns since i dont know yet if i will need them later

## looking for mixed data type columns in the df 

#I didnt find an approach in the last exercises how to find mixed data. Therefore i googled and tried this approach: 

# Check for mixed data types in each column of the DataFrame 'customers'
mixed_types_per_column = cust.applymap(type)

# Identify columns with mixed data types
columns_with_mixed_types = mixed_types_per_column.apply(lambda x: len(set(x))) > 1
columns_with_mixed_types = columns_with_mixed_types[columns_with_mixed_types].index.tolist()

# Display columns with mixed data types
print("Columns with mixed data types:")
print(columns_with_mixed_types)

There are no columns with mixed data type. Could be because I changed some of them beforehand. 

Question: is there an easier or "more intuitive"  approach?

## Finding missing entries

# Count occurrences of each 'User Id'
user_id_counts = cust['User Id'].value_counts(dropna=False)

user_id_counts

# Filter 'User Id's with counts greater than 1 (multiple entries)
multiple_entries = user_id_counts[user_id_counts > 1]

# Display 'User Id's with multiple entries
print("User Ids with multiple entries:")
print(multiple_entries)

now with "First Name"

# Count occurrences of each 'First Name' value 
first_name_counts = cust['First Name'].value_counts(dropna=False)

first_name_counts

-> 11259 Nan entries 

Another approach: 

#count the number of missing (NaN) entries in the 'User Id' column after using value_counts(dropna=False)
#use the isna() method in combination with sum() to count the total number of missing values in that column
missing_values_count = cust['First Name'].isna().sum()

missing_values_count

Another approach for the whole dataframe instead of just the user id or First Name column:

cust.isnull().sum()

-> First Name column has 11259 missing values 
-> I will leave it as it is, because I dont need the first Name yet ( decided later to create a df without missing values for just in case, wont hurt :D ) 

#creating a data frame with only rows where the missing values are located 
df_nan=cust[cust['First Name'].isnull()== True]

Another possibility of treating the missing "First Name" Values would be creating a dataframe without 
missing values. I will do that in case I need it later. 

#creating a df without missing value rows 
cust_no_nan= cust.dropna()

cust_no_nan.shape

cust.shape

206209-194950

-> since the count of rows from the original data frame minus the count of the rows of the missing values data frame
is 11259, the same number like the number of rows with missing entries, the cust_no_nan df is most likely right. 

## Looking for full duplicates

cust_dups = cust[cust.duplicated()]

cust_dups

-> no full duplicates 

# Combining the customer data with the rest of the prepared Instacart data (Task 6) 

Combine your customer data with the rest of your prepared Instacart data. 
(Hint: Make sure the key columns are the same data type!)

#importing the rest of the prepared Instacart data ('ords_prods_merge_4.8.pkl')
ords_prods_merge=pd.read_pickle(r'/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis/02 Data/Prepared Data/ords_prods_merge_4.8.pkl')

#checking the data types 
ords_prods_merge.dtypes

#checking the data types 
cust.dtypes

ords_prods_merge.describe


cust.describe

#in order to have the key columns with the same name I will rename "user_id" to "User Id"
ords_prods_merge.rename(columns={'user_id': 'User Id'}, inplace=True)

#using a inner join to merge the both sets and creating acreate an additional column that indicates the 
#source of the data within that row. 
ords_prods_custs_merge = ords_prods_merge.merge(cust, on = 'User Id', indicator = True)

#checking the new df for match 
ords_prods_custs_merge['_merge'].value_counts()

#finding out how many rows the new df has 
row_count = len(ords_prods_custs_merge)
print(row_count)

# Checking inner merge matches by looking at outer merge
merge_check = ords_prods_merge.merge(cust, on = 'User Id', indicator = True, how = 'outer')

#checking the new df matches 
ords_prods_custs_merge['_merge'].value_counts()

-> Both dataframes have full match. 

## Exporting the data set 

path = '/Users/bente/careerfoundry /exercise 4/Ex.4_25-10-2023 Instacart Basket Analysis'

ords_prods_custs_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_custs_merge.pkl'))
